+++ integration-tests/worker/monorepo/packages/common/index.js
@@ -1,0 +1,8 @@
+function fortyTwo() {
+  return (state) => {
+    state.data = 42;
+    return state;
+  };
+}
+
+module.exports = { fortyTwo };