+++ packages/ws-worker/src/start.ts
@@ -28,7 +28,6 @@ const [minBackoff, maxBackoff] = args.backoff
 
 function engineReady(engine: any) {
   logger.debug('Creating worker instance');
-
   const workerOptions: ServerOptions = {
     port: args.port,
     lightning: args.lightning,
@@ -42,6 +41,9 @@ function engineReady(engine: any) {
     },
     maxWorkflows: args.capacity,
     payloadLimitMb: args.payloadMemory,
+    collectionsVersion: args.collectionsVersion,
+    collectionsUrl: args.collectionsUrl,
+    monorepoDir: args.monorepoDir,
   };
 
   if (args.lightningPublicKey) {